# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Recipez API
  description: |
    REST API for the Recipez recipe management application.

    ## Authentication
    Most endpoints require JWT Bearer token authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```

    ## Authorization Scopes
    Endpoints require specific scopes in the JWT token. Regular users have scopes prefixed with resource types
    (e.g., `recipe:create`, `category:read`). System user (yeschef) has additional administrative scopes.

    ## System User APIs
    Endpoints marked with `[System User]` are restricted to the system user (yeschef) and are used for
    authentication flows and user management. These require system-level scopes.
  version: 1.0.0
  contact:
    name: Recipez Support
  license:
    name: Proprietary
servers:
  - url: /
    description: Current server
...
paths:
...
  /api/recipe/create:
    post:
      tags:
        - Recipe
      summary: Create a new recipe
      description: Create a new recipe. Requires `recipe:create` scope.
      operationId: createRecipe
      security:
        - bearerAuth: [recipe:create]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRecipeRequest'
      responses:
        '200':
          description: Recipe created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Recipe with this name already exists
        '500':
          $ref: '#/components/responses/InternalError'
  /api/recipe/all:
    get:
      tags:
        - Recipe
      summary: Get all recipes
      description: |
        Retrieve all recipes with nested category, image, author, ingredients, and steps.
        Requires `recipe:read` scope.
      operationId: getAllRecipes
      security:
        - bearerAuth: [recipe:read]
      responses:
        '200':
          description: List of all recipes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipesListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/recipe/{recipeId}:
    get:
      tags:
        - Recipe
      summary: Get a single recipe
      description: |
        Retrieve a single recipe by ID with nested category, image, author, ingredients, and steps.
        Requires `recipe:read` scope.
      operationId: getRecipe
      security:
        - bearerAuth: [recipe:read]
      parameters:
        - $ref: '#/components/parameters/RecipeId'
      responses:
        '200':
          description: Recipe retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeDetailResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/recipe/update/{recipeId}:
    put:
      tags:
        - Recipe
      summary: Update a recipe
      description: |
        Update an existing recipe. Requires `recipe:update` scope and recipe ownership.
        Only the recipe author can update the recipe.
      operationId: updateRecipe
      security:
        - bearerAuth: [recipe:update]
      parameters:
        - $ref: '#/components/parameters/RecipeId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRecipeRequest'
      responses:
        '200':
          description: Recipe updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: User is not the recipe author
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/recipe/delete/{recipeId}:
    delete:
      tags:
        - Recipe
      summary: Delete a recipe
      description: |
        Delete a recipe. Requires `recipe:delete` scope and recipe ownership.
        Also cleans up associated image if not a default image.
      operationId: deleteRecipe
      security:
        - bearerAuth: [recipe:delete]
      parameters:
        - $ref: '#/components/parameters/RecipeId'
      responses:
        '200':
          description: Recipe deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Invalid recipe ID
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: User is not the recipe author
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/recipe/batch-update-category:
    post:
      tags:
        - Recipe
      summary: Batch update recipe categories
      description: |
        Update categories for multiple recipes at once. Requires `recipe:update` scope.
        Only updates recipes owned by the current user.
      operationId: batchUpdateRecipeCategories
      security:
        - bearerAuth: [recipe:update]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchUpdateCategoryRequest'
      responses:
        '200':
          description: Batch update completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchUpdateCategoryResponse'
        '400':
          description: No updates provided
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'

  # ==================== CATEGORY ENDPOINTS ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token with scopes. Include in Authorization header as: `Bearer <token>`

        Token contains:
        - `sub`: User subject ID
        - `scope`: Space-separated list of permission scopes
        - `iat`: Issued at timestamp
        - `exp`: Expiration time (12 hours from issue)
        - `iss`: Token issuer
        - `aud`: Token audience
  parameters:
    RecipeId:
      name: recipeId
      in: path
      required: true
      description: Recipe UUID
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Bad request - invalid input data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized - missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden - insufficient permissions or scope
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        response:
          type: object
          properties:
            error:
              type: string
              description: Error message
    SuccessResponse:
      type: object
      properties:
        response:
          type: object
          properties:
            success:
              type: boolean
              example: true
...
    User:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
        user_sub:
          type: string
          format: uuid
          description: Subject ID used in JWT
        user_email:
          type: string
          description: Encrypted email address
        user_name:
          type: string
        user_created_at:
          type: string
          format: date-time
          nullable: true
        user_profile_image_url:
          type: string
          nullable: true
...
    Recipe:
      type: object
      properties:
        recipe_id:
          type: string
          format: uuid
        recipe_name:
          type: string
        recipe_description:
          type: string
        recipe_category_id:
          type: string
          format: uuid
        recipe_image_id:
          type: string
          format: uuid
          nullable: true
        recipe_author_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
    RecipeWithRelations:
      allOf:
        - $ref: '#/components/schemas/Recipe'
        - type: object
          properties:
            recipe_category:
              $ref: '#/components/schemas/Category'
            recipe_image:
              $ref: '#/components/schemas/Image'
            recipe_author:
              $ref: '#/components/schemas/User'
            author:
              $ref: '#/components/schemas/User'
              description: Alias for recipe_author
            recipe_ingredients:
              type: array
              items:
                $ref: '#/components/schemas/Ingredient'
            recipe_steps:
              type: array
              items:
                $ref: '#/components/schemas/Step'
    CreateRecipeRequest:
      type: object
      required:
        - recipe_name
        - recipe_description
        - recipe_category_id
        - recipe_image_id
        - recipe_author_id
      properties:
        recipe_name:
          type: string
          minLength: 2
          maxLength: 50
          pattern: "^[0-9a-zA-Z-_()., '&]+$"
        recipe_description:
          type: string
          minLength: 2
          maxLength: 500
        recipe_category_id:
          type: string
          format: uuid
        recipe_image_id:
          type: string
          format: uuid
        recipe_author_id:
          type: string
          format: uuid
    UpdateRecipeRequest:
      type: object
      properties:
        recipe_name:
          type: string
          minLength: 2
          maxLength: 50
          pattern: "^[0-9a-zA-Z-_()., '&]+$"
          nullable: true
        recipe_description:
          type: string
          minLength: 2
          maxLength: 500
          nullable: true
        recipe_category_id:
          type: string
          format: uuid
          nullable: true
        recipe_image_id:
          type: string
          format: uuid
          nullable: true
    RecipeResponse:
      type: object
      properties:
        response:
          type: object
          properties:
            recipe:
              $ref: '#/components/schemas/Recipe'
    RecipeDetailResponse:
      type: object
      properties:
        response:
          $ref: '#/components/schemas/RecipeWithRelations'
    RecipesListResponse:
      type: object
      properties:
        response:
          type: array
          items:
            $ref: '#/components/schemas/RecipeWithRelations'
    BatchUpdateCategoryRequest:
      type: object
      required:
        - updates
      properties:
        updates:
          type: array
          items:
            type: object
            required:
              - recipe_id
              - category_id
            properties:
              recipe_id:
                type: string
                format: uuid
              category_id:
                type: string
                format: uuid
    BatchUpdateCategoryResponse:
      type: object
      properties:
        response:
          type: object
          properties:
            success:
              type: boolean
            updated_count:
              type: integer
            failed_count:
              type: integer
            updated:
              type: array
              items:
                type: object
                properties:
                  recipe_id:
                    type: string
                    format: uuid
                  recipe_name:
                    type: string
                  new_category_id:
                    type: string
                    format: uuid
                  new_category_name:
                    type: string
            failed:
              type: array
              items:
                type: object
                properties:
                  recipe_id:
                    type: string
                    format: uuid
                  reason:
                    type: string

    # ==================== CATEGORY SCHEMAS ====================
    Category:
      type: object
      properties:
        category_id:
          type: string
          format: uuid
        category_name:
          type: string
        category_author_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
...
    Ingredient:
      type: object
      properties:
        ingredient_id:
          type: string
          format: uuid
        ingredient_name:
          type: string
        ingredient_quantity:
          type: string
        ingredient_measurement:
          type: string
        ingredient_author_id:
          type: string
          format: uuid
        ingredient_recipe_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
...
    Step:
      type: object
      properties:
        step_id:
          type: string
          format: uuid
        step_text:
          type: string
        step_author_id:
          type: string
          format: uuid
        step_recipe_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
...
    Image:
      type: object
      properties:
        image_id:
          type: string
          format: uuid
        image_url:
          type: string
        image_author_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
...